feat(app): dynamic authentication provider support
This change adds support for loading authentication providers or modules
from dynamic plugins.  An environment variable
ENABLE_AUTH_PROVIDER_MODULE_OVERRIDE controls whether or not the backend
installs the default authentication provider module.  When this override
is enabled dynamic plugins can be used to supply custom authentication
providers.  This change also modifies the frontend to accept a
SignInPage component from a dynamic plugin.  The current method of
discovery is to scan all of the loaded dynamic plugin modules for an
export named "SignInPage", when found it will be used instead of the
default SignInPage.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>